Make core updates dismissible (#198)
Some people don't want to upgrade to the local package because they have more than one locale on the installation or because they need to wait a few days until the local package is released.

Let's re-add the possibility to dismiss these updates and handle situations with multiple available core updates more elegantly.

This change allows showing multiple updates in the list table and ensures that only 1 core update can be run at the same time. When clicking "Update All", it will update the localized package.

Also, this change prevents showing an AYS dialog when the core update is finished before a plugin update by storing the redirect URL and waiting until the queue is empty.