Improve position of the inline error notice for plugin delete failures (#239)
Previously, the error notice was added before the plugin description. Now it is added in a separate row after the plugin, just like the update notices.

If such an update row already exists (e.g. if there's a plugin update available or there was an error message before), it adds the notice to that row, overriding any previous error message.